refactor(quiz): improve file naming and calculation logic
- Rename UI components from PascalCase to kebab-case
- Fix quiz score calculation using Object.entries instead of Object.values
- Remove unnecessary dynamic import for ResultChart component
- Simplify image capture logic by removing border width calculations
- Remove unused priority prop from Image component
