+++ entities/quiz/ui/answer-option.tsx
