+++ entities/quiz/ui/index.ts
@@ -1,4 +1,4 @@
-export * from './QuestionCard';
-export * from './AnswerOption';
-export * from './ProgressBar';
-export * from './ResultCard';
+export * from './question-card';
+export * from './answer-option';
+export * from './progress-bar';
+export * from './result-card';
