+++ entities/quiz/ui/progress-bar.tsx
