+++ entities/quiz/ui/result-card.tsx
